Added doc about Singularity, first workflow draft
We will need to experiment with the --sandbox option in order
to run UFS SRW as a Singularity container - the reasoning for
this has been added to ./containers/README.md.

A first draft of the workflow orchestration files for running
UFS SRW on PW has been added in ./workflow.  Currently the
workflow launches, but run to completion has not been verified
and significant testing is required.