Adjusted workflow for concurrent runs
Changed the workflow config and passing ENV vars
to container so that workflow runs in silo (e.g. mktemp)
directories.  This will allow for simultaneous runs
on the same file system without race conditions/clashed
and/or easy direct copy back to a single directory.

Ultimately, the goal is to pass this workflow a .setup
and a .config file, with each controlling the data access
and the model run, respectively.  This allows the user
to try a range of different models as well as different cloud data
access options. Also, since all runs are siloed, it should
be straightforward to test simultaneous repetitions of
a given model run to see how data access scales.