Added head node setup script
The head node needs to have a Conda environment for running
remote ipykernels.  This simple script is a quasi-automation
of this process.  I recommend running this in /var/lib/pworks
on the HEAD node, and not in $HOME as some cluster workers
attempt to chown -r/chmod -r $HOME which takes a very long
time if a miniconda dir is in $HOME b/c there are loads of files.